requirements.txt
streamlit
google-generativeai
PyMuPDF
requests
textstat
python-dotenv
gTTS
